Event consumption/reducer interface are too bare
It makes sense to leave the method of creating a reducer to the user
providing `by-name` as the dispatch function. That can be further
simplified to a macro for creating the entire `defmulti`. Also the
event-consumer function could easily be replaced by a macro that
allows for more flexibility when dealing with events by composing
`contrib.core/dochan`